refactor: Client -> HttpClient
